onnx_requirements.txt
tensorflow
tf-keras
onnx
onnxruntime

# Need to edit utils from np.object to object
git+https://github.com/onnx/tensorflow-onnx

jax[cuda12]

mujoco
mujoco-mjx
flax
optax
brax
distrax
orbax-checkpoint